divided pinger
