+++ docker-compose.yaml
@@ -82,7 +82,7 @@ services:
     image: provectuslabs/kafka-ui
     container_name: kafka-ui
     ports:
-      - "8090:8080"
+      - "8090:8090"
     restart: always
     environment:
       - KAFKA_CLUSTERS_0_NAME=local
