+++ pinger/cmd/app/config/config.go
@@ -7,7 +7,6 @@ import (
 	"io/ioutil"
 )
 
-// Config содержит конфигурацию приложения
 type Config struct {
 	Kafka struct {
 		Broker string `yaml:"broker"`
@@ -17,17 +16,14 @@ type Config struct {
 	} `yaml:"ping"`
 }
 
-// Load загружает конфигурацию из YAML файла
 func Load() *Config {
 	cfg := &Config{}
 	
-	// Чтение файла конфигурации
 	data, err := ioutil.ReadFile("config.yaml")
 	if err != nil {
 		panic(err)
 	}
 
-	// Парсинг YAML
 	err = yaml.Unmarshal(data, cfg)
 	if err != nil {
 		panic(err)
@@ -36,7 +32,6 @@ func Load() *Config {
 	return cfg
 }
 
-// GetKafkaBroker возвращает адрес брокера Kafka
 func (c *Config) GetKafkaBroker() string {
 	if broker := os.Getenv("KAFKA_BROKER"); broker != "" {
 		return broker
@@ -44,7 +39,6 @@ func (c *Config) GetKafkaBroker() string {
 	return c.Kafka.Broker
 }
 
-// GetPingInterval возвращает интервал ping-запросов
 func (c *Config) GetPingInterval() time.Duration {
 	if interval := os.Getenv("PING_INTERVAL"); interval != "" {
 		duration, err := time.ParseDuration(interval)
