+++ pinger/domain/repository/docker.go
@@ -7,4 +7,5 @@ import (
 
 type DockerRepository interface {
     GetContainerInfos(ctx context.Context) (map[string]model.ContainerInfo, error)
+	Close() error
 }
