+++ pinger/domain/repository/ping.go
