+++ pinger/domain/service/docker.go
@@ -1,0 +1,12 @@
+package service
+
+import (
+	"context"
+	"github.com/docker/docker/api/types/container"
+	"github.com/docker/docker/api/types"
+)
+
+type DockerClient interface {
+	ContainerList(ctx context.Context, options container.ListOptions) ([]types.Container, error)
+	Close() error
+}
