+++ pinger/internal/ping/ping.go
@@ -8,26 +8,8 @@ import (
 
 	probing "github.com/prometheus-community/pro-bing"
 	"pinger/domain/model"
-	"pinger/domain/repository"
 )
 
-type Service struct {
-	dockerRepo repository.DockerRepository
-	pingRepo   repository.PingRepository
-	interval   time.Duration
-}
-
-func New(
-	dockerRepo repository.DockerRepository,
-	pingRepo repository.PingRepository,
-	interval time.Duration,
-) *Service {
-	return &Service{
-		dockerRepo: dockerRepo,
-		pingRepo:   pingRepo,
-		interval:   interval,
-	}
-}
 
 func (s *Service) Run(ctx context.Context) {
 	ticker := time.NewTicker(s.interval)
