+++ pinger/service/docker/docker.go
@@ -6,63 +6,69 @@ import (
 
 	"pinger/domain/model"
 	"pinger/domain/repository"
+	"pinger/domain/service"
 
 	"github.com/docker/docker/api/types"
 	"github.com/docker/docker/api/types/container"
 	dockerClient "github.com/docker/docker/client"
 )
 
-type Client struct {
-    cli DockerInterface
-}
-
-type DockerInterface interface {
-	ContainerList(ctx context.Context, options container.ListOptions) ([]types.Container, error)
-	Close() error
+type dockerRepository struct {
+	client service.DockerClient
 }
 
 func New() (repository.DockerRepository, error) {
-    cli, err := dockerClient.NewClientWithOpts(
-        dockerClient.FromEnv,
-        dockerClient.WithVersion("1.47"),
-    )
-    if err != nil {
-        return nil, err
-    }
+	cli, err := dockerClient.NewClientWithOpts(
+		dockerClient.FromEnv,
+		dockerClient.WithVersion("1.47"),
+	)
+	if err != nil {
+		return nil, err
+	}
 
-    return &Client{cli: cli}, nil
+	return &dockerRepository{
+		client: cli,
+	}, nil
 }
 
-func (c *Client) GetContainerInfos(ctx context.Context) (map[string]model.ContainerInfo, error) {
-    containers, err := c.cli.ContainerList(ctx, container.ListOptions{})
-    if err != nil {
-        return nil, err
-    }
+func (r *dockerRepository) GetContainerInfos(ctx context.Context) (map[string]model.ContainerInfo, error) {
+	containers, err := r.client.ContainerList(ctx, container.ListOptions{})
+	if err != nil {
+		return nil, err
+	}
 
-    results := make(map[string]model.ContainerInfo)
-    for _, cntr := range containers {
-        var name string
-        if len(cntr.Names) > 0 {
-            name = strings.TrimPrefix(cntr.Names[0], "/")
-        } else {
-            name = cntr.ID[:12]
-        }
+	results := make(map[string]model.ContainerInfo)
+	for _, cntr := range containers {
+		results[cntr.ID] = convertToContainerInfo(cntr)
+	}
 
-        var ips []string
-        if cntr.NetworkSettings != nil {
-            for _, netSettings := range cntr.NetworkSettings.Networks {
-                if netSettings.IPAddress != "" {
-                    ips = append(ips, netSettings.IPAddress)
-                }
-            }
-        }
+	return results, nil
+}
 
-        results[cntr.ID] = model.ContainerInfo{
-            ID:   cntr.ID,
-            Name: name,
-            IPs:  ips,
-        }
-    }
+func (r *dockerRepository) Close() error {
+	return r.client.Close()
+}
 
-    return results, nil
+func convertToContainerInfo(cntr types.Container) model.ContainerInfo {
+	var name string
+	if len(cntr.Names) > 0 {
+		name = strings.TrimPrefix(cntr.Names[0], "/")
+	} else {
+		name = cntr.ID[:12]
+	}
+
+	var ips []string
+	if cntr.NetworkSettings != nil {
+		for _, netSettings := range cntr.NetworkSettings.Networks {
+			if netSettings.IPAddress != "" {
+				ips = append(ips, netSettings.IPAddress)
+			}
+		}
+	}
+
+	return model.ContainerInfo{
+		ID:   cntr.ID,
+		Name: name,
+		IPs:  ips,
+	}
 }
